chore: move event hook and sidebar to the local-chart folder
